# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.BankResponse:
    properties:
      bank_type:
        example: theory
        type: string
      category_id:
        example: a1b2c3d4e5f6g7h8
        type: string
      difficulty:
        example: medium
        type: string
      id:
        example: x9y8z7w6v5u4t3s2
        type: string
      language:
        example: go
        type: string
      mastery:
        example: 42
        type: integer
      subject:
        example: Go concurrency patterns
        type: string
    type: object
...
  api.CategoryResponse:
    properties:
      folder_id:
        example: f1o2l3d4e5r6i7d8
        type: string
      id:
        example: a1b2c3d4e5f6g7h8
        type: string
      mastery:
        example: 42
        type: integer
      name:
        example: Golang
        type: string
      sort_order:
        example: 0
        type: integer
    type: object
  api.CategoryStatsResponse:
    properties:
      category_id:
        example: a1b2c3d4e5f6g7h8
        type: string
      mastery:
        example: 42
        type: integer
    type: object
...
  api.CreateCategoryRequest:
    properties:
      folder_id:
        example: f1o2l3d4e5r6i7d8
        type: string
      name:
        example: Golang
        type: string
    type: object
...
  api.GetCategoryResponse:
    properties:
      banks:
        items:
          $ref: '#/definitions/api.BankResponse'
        type: array
      folder_id:
        example: f1o2l3d4e5r6i7d8
        type: string
      id:
        example: a1b2c3d4e5f6g7h8
        type: string
      mastery:
        example: 42
        type: integer
      name:
        example: Golang
        type: string
    type: object
...
  api.ReorderCategoriesRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
...
  api.UpdateCategoryFolderRequest:
    properties:
      folder_id:
        example: f1o2l3d4e5r6i7d8
        type: string
    type: object
  api.UpdateCategoryRequest:
    properties:
      name:
        example: Rust
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Personal learning companion — create question banks, practice, and
    let AI grade your answers.
  title: Remaimber-it API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Returns all categories with their mastery scores.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.CategoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category for organizing question banks. Optionally
        assign to a folder.
      parameters:
      - description: Category to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a category
      tags:
      - Categories
  /categories/{categoryID}:
    delete:
      description: Delete a category and cascade-delete all its banks and questions.
      parameters:
      - description: Category ID
        in: path
        name: categoryID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a category
      tags:
      - Categories
    get:
      description: Returns a category with all its question banks.
      parameters:
      - description: Category ID
        in: path
        name: categoryID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetCategoryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update the name of an existing category.
      parameters:
      - description: Category ID
        in: path
        name: categoryID
        required: true
        type: string
      - description: New category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a category
      tags:
      - Categories
  /categories/{categoryID}/banks:
    get:
      description: Returns all question banks belonging to a category.
      parameters:
      - description: Category ID
        in: path
        name: categoryID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.BankResponse'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List banks by category
      tags:
      - Categories
  /categories/{categoryID}/folder:
    patch:
      consumes:
      - application/json
      description: Move a category to a different folder (or set to null to unfiled).
      parameters:
      - description: Category ID
        in: path
        name: categoryID
        required: true
        type: string
      - description: New folder
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateCategoryFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update category folder
      tags:
      - Categories
  /categories/{categoryID}/stats:
    get:
      description: Returns the aggregate mastery score for a category.
      parameters:
      - description: Category ID
        in: path
        name: categoryID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CategoryStatsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category stats
      tags:
      - Categories
  /categories/reorder:
    patch:
      consumes:
      - application/json
      description: Rewrites sort_order for the given categories to match the order
        of the ids array.
      parameters:
      - description: Category IDs in the desired order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ReorderCategoriesRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reorder categories
      tags:
      - Categories
...
swagger: "2.0"